Fixes and improvements for flatpak processing

+++ files/dotfiles/Configs/scripts/fix-vaapi-chromium-flatpak.sh
@@ -12,6 +12,18 @@ EOF
 	exit 1
 }
 
+do_overrides() {
+	local app=$1
+	local dri=$2
+	local user_opt=$3
+
+	flatpak override "$user_opt" --reset "$APP" 2>/dev/null
+	flatpak override "$user_opt" \
+		--env=LIBVA_DRIVER_NAME=nvidia \
+		--env=LIBVA_DRIVERS_PATH="$dri" \
+		--env=NVD_BACKEND=direct "$app"
+}
+
 USER_FLAG=false
 APP_NAME=""
 LOCAL_PATH=""
@@ -58,13 +70,8 @@ done
 APP="${APP_NAME:-$DEFAULT_APP}"
 LOCAL_DRI_PATH="${LOCAL_PATH:-/usr/lib64/dri/nvidia_drv_video.so}"
 BASE_DIR="$(realpath "$HOME")/.var/app/$APP"
-RUNFILE="$BASE_DIR/.chrome-vaapi-fix-applied"
-if [ -e "$RUNFILE" ]; then
-	echo "Detected '$RUNFILE', fix has probably been applied, aborting..."
-	exit 1
-fi
-
 DRI_PATH="$BASE_DIR/dri"
+REMOTE_DRI_PATH="$DRI_PATH/nvidia_drv_video.so"
 user_opt="${USER_FLAG:+--user}"
 
 [ -z "$APP_NAME" ] && APP_NAME=$APP
@@ -74,18 +81,18 @@ if ! (flatpak "$user_opt" list | grep -q "$APP"); then
 	exit 1
 fi
 
+if [ -r "$REMOTE_DRI_PATH" ]; then
+	do_overrides "$APP" "$REMOTE_DRI_PATH" "$user_opt"
+	echo "Detected '$REMOTE_DRI_PATH', fix has already run, redoing overrides..."
+	exit 0
+fi
+
 [ ! -r "$LOCAL_DRI_PATH" ] && echo "Can't find driver at $LOCAL_DRI_PATH" && exit 1
 
 rm -rf "$DRI_PATH" &&
 	mkdir -p "$DRI_PATH" &&
 	cp -f "$LOCAL_DRI_PATH" "$DRI_PATH/nvidia_drv_video.so"
 
-flatpak override "$user_opt" --reset "$APP" 2>/dev/null
-flatpak override "$user_opt" \
-	--env=LIBVA_DRIVER_NAME=nvidia \
-	--env=LIBVA_DRIVERS_PATH="$DRI_PATH" \
-	--env=NVD_BACKEND=direct "$APP"
-
-mkdir -p "$BASE_DIR" && touch "$RUNFILE"
+do_overrides "$APP" "$REMOTE_DRI_PATH" "$user_opt"
 
 echo "NVIDIA driver configured for $APP at $DRI_PATH"
